Adding loggging for the debug process
