+++ src/pyoslc_server/utils.py
@@ -1,0 +1,20 @@
+from .http import HTTPStatus
+
+
+def unpack(response, default_code=HTTPStatus.OK):
+    if not isinstance(response, tuple):
+        # data only
+        return response, default_code, {}
+    elif len(response) == 1:
+        # data only as tuple
+        return response[0], default_code, {}
+    elif len(response) == 2:
+        # data and code
+        data, code = response
+        return data, code, {}
+    elif len(response) == 3:
+        # data, code and headers
+        data, code, headers = response
+        return data, code or default_code, headers
+    else:
+        raise ValueError("Too many response values")
